Update to Bitcoin Classic 1.3.2. Move to Alpine 3.6.
Updated to Bitcoin Classic 1.3.2.

Switched from Alpine 3.5 to 3.6. Also changed the GPG keyserver to their ipv4
pool. This is to get around an issue with Docker/GPG where it would often times
fail on lookup, returning "Address not available". Some reports pointed to an
issue where it would try to use IPv6 even when it wasn't available.